Refactor authentication to use UserContext for login and signup functionality
